+++ DAO/artwork_Service.py
@@ -47,33 +47,47 @@ class ArtworkService(IArtworkService,DBConnection):
 
     def addArtwork(self,new_artwork):
         try:
-            self.cursor.execute("insert into Artwork(artworkID,description,title,creationDate,medium,imageURL,artistID) VALUES(?,?,?,?,?,?,?)",
-                      (new_artwork.artworkID,new_artwork.description,new_artwork.title,new_artwork.creationDate,new_artwork.medium,new_artwork.imageURL,new_artwork.artistID)
+            self.cursor.execute("insert into Artwork(title,description,creationDate,medium,imageURL) OUTPUT inserted.artworkID VALUES(?,?,?,?,?)",
+                      (new_artwork.title,new_artwork.description,new_artwork.creationDate,new_artwork.medium,new_artwork.imageURL)
                         )
-
+            new_artwork_id = self.cursor.fetchone()[0]
             self.conn.commit() 
-            return new_artwork.artworkID
+            return new_artwork_id
         except Exception as e:
             print("Error!!",e)
             return None
         
     def removeArtwork(self,artworkId):
         try:
-            self.cursor.execute("Delete FROM User_Favorite_Artwork WHERE artworkId = ?", (artworkId,))
-            self.cursor.execute("Delete FROM Artwork_Gallery WHERE artworkId = ?", (artworkId,))
-            self.cursor.execute("Delete from artwork where artworkId=?",(artworkId))
-           
-            self.conn.commit()
+            self.cursor.execute("SELECT * FROM Artwork WHERE artworkID = ?", (artworkId,))
+            artwork = self.cursor.fetchone()
+            if artwork is None:
+                raise ArtworkNotFoundException(artworkId)
+            else:
+                self.cursor.execute("Delete FROM User_Favorite_Artwork WHERE artworkId = ?", (artworkId,))
+                self.cursor.execute("Delete FROM Artwork_Gallery WHERE artworkId = ?", (artworkId,))
+                self.cursor.execute("Delete from artwork where artworkId=?",(artworkId))
+                self.conn.commit()
+        except ArtworkNotFoundException as e:
+            print("Error !!!",e)
         except Exception as e:
-            print(e)
+            print("Error !!",e)
        
-    def updateArtwork(self,artworkId,description,title,creationDate,medium,imageURL,artistID):
+    def updateArtwork(self,artworkId,title,description,creationDate,medium,imageURL):
         try:
-            self.cursor.execute("update artwork SET description = ?, title = ?, creationDate = ?, medium = ?, imageURL = ?, artistID = ? WHERE artworkId = ? ",
-                        (description,title,creationDate,medium,imageURL,artistID,artworkId)
+            self.cursor.execute("SELECT * FROM Artwork WHERE artworkID = ?", (artworkId,))
+            artwork = self.cursor.fetchone()
+        
+            if artwork is None:
+                raise ArtworkNotFoundException(artworkId)
+            else:
+                self.cursor.execute("update artwork SET  title = ?, description = ?, creationDate = ?, medium = ?, imageURL = ? WHERE artworkId = ? ",
+                        (title,description,creationDate,medium,imageURL,artworkId)
                         )
             self.conn.commit()
             return True
+        except ArtworkNotFoundException as e:
+            print("Error !!!",e)
         except Exception as e:
             print(e)
             return False
